Don't bootstrap logging when using ResourceExtension
The reason is that, when ResourceExtension bootstraps logging, the
log level is set to WARN and then we never see logging at lower levels
once the resource test has completed. By telling ResourceExtension not
to bootstrap logging, it won't change our test logging configuration.
This should really be the default behavior IMHO...