Enhance logging in ApplicationErrorVerificationsTest
Enhance logging in the BusinessLogicService test class to log at
WARN level if we threw an ApplicationError, and INFO level if there
was no ApplicationError.

@implNote:
Because SLF4J doesn't have have a generic log method that accepts an
SLF4J Level - even in the new 2.0.0 API, see long thread here:
https://jira.qos.ch/browse/SLF4J-124 - I resorted to defining my own
small crappy enum and method to do this.