Use KiwiInternetAddresses to get the InetAddress for localhost
This removes boilerplate and also eliminates the catch of the
UnknownHostException which was not covered by tests (since it's not
trivial to force the JDK InetAddress#getLocalHost to throw an
UnknownHostException). Technically the orElseGet still isn't covered,
but it's not worth bothering with since it is so unlikely to ever
happen.

Closes #217